app completed

+++ zrmgz/Model/Account.swift
@@ -10,13 +10,13 @@ import Foundation
 struct Account {
     
     var id: Int64
-    var sortId: Int64
+    var sortId: Int
     var sortName: String
     var type: Int
     var num: Double
     var year: Int
     var month: Int
     var day: Int
-    var date: Int64
+    var date: Int
     
 }
